Bit of a refactor to separate model file, some midi utils for effect, concrete midi effects to own file
